create_index.py no longer relies on importer settings
