+++ labonneboite/common/util.py
@@ -4,6 +4,7 @@ from urlparse import urlparse
 import logging
 import unicodedata
 import urllib
+from backports.functools_lru_cache import lru_cache
 
 from flask import request
 
@@ -12,6 +13,16 @@ from labonneboite.common.load_data import load_contact_modes
 
 logger = logging.getLogger('main')
 
+@lru_cache(maxsize=None)
+def get_departements(largest_ones_first=False):
+    departements = ["{:02d}".format(d) for d in range(1, 96)] + ['97']
+    if largest_ones_first:
+        departements.remove('75')
+        departements[:0] = ['75']
+    return departements
+
+DEPARTEMENTS = get_departements()
+DEPARTEMENTS_WITH_LARGEST_ONES_FIRST = get_departements(largest_ones_first=True)
 
 def get_search_url(base_url, request_args, naf=None):
     query_string = {}
